Character detail

+++ backend/dna-tools-api/src/main/java/com/dna/tools/domain/character/dto/CharacterStatsResponse.java
@@ -7,14 +7,14 @@ import lombok.Getter;
 @Getter
 public class CharacterStatsResponse {
 
-    private final int attack;
+    private final BigDecimal attack;
     private final int hp;
     private final int defense;
     private final int maxMentality;
     private final BigDecimal resolve;
     private final BigDecimal morale;
 
-    public CharacterStatsResponse(int attack, int hp, int defense, int maxMentality, BigDecimal resolve,
+    public CharacterStatsResponse(BigDecimal attack, int hp, int defense, int maxMentality, BigDecimal resolve,
             BigDecimal morale) {
         this.attack = attack;
         this.hp = hp;
